# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ludus UX API
  description: |
    REST API for the Ludus UX web application — a Next.js frontend for managing
    [Ludus](https://ludus.cloud) cyber range infrastructure, including VM ranges,
    content sources (sync / force re-install), templates, user management,
    VNC/SPICE console access, GOAD (Game of Active Directory) lab orchestration,
    and LudusHound BloodHound→Ludus deploys.

    ## Authentication

    All protected endpoints require a valid session cookie set by `POST /api/auth/login`.
    The cookie is AES-256-GCM encrypted and HTTP-only. Sessions expire after 24 hours.

    **Admin impersonation headers** (admin sessions only):
    - `X-Impersonate-As` — target username
    - `X-Impersonate-Apikey` — target user's Ludus API key
    - `X-Ludus-Admin: true` — route to admin API port (8081)
    - `X-Ludus-User` — override Ludus user context
  # Keep in sync with package.json / APP_VERSION (src/lib/changelog.ts).
  version: 1.4.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
    description: Ludus UX application server
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ludus_session
      description: AES-256-GCM encrypted HTTP-only session cookie set by the login endpoint.
  parameters:
    impersonateAs:
      name: X-Impersonate-As
      in: header
      required: false
      schema:
        type: string
      description: Admin only. Scope the request to this username.
    impersonateApikey:
      name: X-Impersonate-Apikey
      in: header
      required: false
      schema:
        type: string
      description: Admin only. Use this user's Ludus API key instead of the session key.
...
  schemas:
...
    RuntimeSettings:
      type: object
      properties:
        ludusUrl:
          type: string
          format: uri
          example: https://192.168.1.10:8080
        ludusAdminUrl:
          type: string
          format: uri
          example: https://192.168.1.10:8081
        sshHost:
          type: string
        sshPort:
          type: integer
          default: 22
        goadPath:
          type: string
          example: /opt/GOAD
        ludusInstallPath:
          type: string
          example: /opt/ludus
        rootApiKey:
          type: string
        proxmoxSshUser:
          type: string
          example: root
        proxmoxSshPassword:
          type: string
          format: password

    # ---- Range / VM ----
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
  responses:
    Unauthorized:
      description: No valid session cookie or session has expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
    Forbidden:
      description: Authenticated but insufficient privileges (admin required).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Forbidden

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/settings:
    get:
      tags: [Settings]
      summary: Get runtime settings
      description: Returns the current runtime configuration (Ludus URL, SSH config, GOAD path, etc.).
      operationId: getSettings
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Current settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Settings]
      summary: Update runtime settings
      description: Persists updated runtime settings to SQLite. Admin required.
      operationId: updateSettings
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ludusUrl:
                  type: string
                  format: uri
                ludusAdminUrl:
                  type: string
                  format: uri
                sshHost:
                  type: string
                sshPort:
                  type: integer
                goadPath:
                  type: string
                ludusInstallPath:
                  type: string
                proxmoxSshUser:
                  type: string
                proxmoxSshPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Updated settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/goad/instances/{instanceId}/init-range:
    post:
      tags: [GOAD]
      summary: Ensure GOAD instance has a Ludus range
      description: |
        Idempotent. Creates a Ludus range for the GOAD instance if one does
        not already exist, assigns it to the user, and writes the `rangeID`
        to the instance workspace file. Returns the existing `rangeId` if
        already set.
      operationId: initGoadInstanceRange
      security:
        - sessionCookie: []
      parameters:
        - name: instanceId
          in: path
          required: true
          schema:
            type: string
          description: GOAD instance directory name.
        - $ref: '#/components/parameters/impersonateAs'
        - $ref: '#/components/parameters/impersonateApikey'
      responses:
        '200':
          description: Range ID (existing or newly created).
          content:
            application/json:
              schema:
                type: object
                properties:
                  rangeId:
                    type: string
                  created:
                    type: boolean
                    description: "`true` if a new range was created, `false` if it already existed."
                required: [rangeId, created]
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/goad/instances/{instanceId}/sync-ips:
    post:
      tags: [GOAD]
      summary: Sync GOAD inventory with Ludus IP range
      description: |
        Syncs GOAD Ansible inventory files with the actual Ludus-assigned IP
        range. Replaces stale placeholder IPs in `instance.json` and all
        inventory files via SSH. Calls the Ludus API to resolve the
        `rangeNumber`.
      operationId: syncGoadIps
      security:
        - sessionCookie: []
      parameters:
        - name: instanceId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/impersonateApikey'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ludusRangeId:
                  type: string
      responses:
        '200':
          description: Sync result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  oldIpRange:
                    type: string
                    example: 10.4.10
                  newIpRange:
                    type: string
                    example: 10.5.10
                  rangeNumber:
                    type: integer
                  updates:
                    type: array
                    items:
                      type: string
                    description: List of files that were updated.
                  errors:
                    type: array
                    items:
                      type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/goad/tasks/{taskId}/stop:
    post:
      tags: [GOAD]
      summary: Stop a running GOAD task
      description: |
        Kills the in-flight SSH/Ansible process (sends Ctrl+C to the PTY) and
        marks the task as `aborted`. Works even after the original SSE client
        disconnected.
      operationId: stopGoadTask
      security:
        - sessionCookie: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Stop result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  killed:
                    type: boolean
                    description: Whether the underlying process was actually killed.
                required: [success, killed]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Task not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'